+++ todos/create.py
@@ -5,6 +5,8 @@ import time
 import uuid
 
 import boto3
+from todoTableClass import handler as todoTableClass
+
 dynamodb = boto3.resource('dynamodb')
 
 
@@ -14,20 +16,22 @@ def create(event, context):
         logging.error("Validation Failed")
         raise Exception("Couldn't create the todo item.")
     
-    timestamp = str(time.time())
+    tdCreate = todoTableClass(table = os.environ['DYNAMODB_TABLE'], dynamodb = dynamodb)
+    item = tdCreate.put_todo(data['text'])
+    #timestamp = str(time.time())
 
-    table = dynamodb.Table(os.environ['DYNAMODB_TABLE'])
+    #table = dynamodb.Table(os.environ['DYNAMODB_TABLE'])
 
-    item = {
-        'id': str(uuid.uuid1()),
-        'text': data['text'],
-        'checked': False,
-        'createdAt': timestamp,
-        'updatedAt': timestamp,
-    }
+    #item = {
+    #    'id': str(uuid.uuid1()),
+    #    'text': data['text'],
+    #    'checked': False,
+    #    'createdAt': timestamp,
+    #    'updatedAt': timestamp,
+    #}
 
     # write the todo to the database
-    table.put_item(Item=item)
+    #table.put_item(Item=item)
 
     # create a response
     response = {