+++ todos/get.py
@@ -1,26 +1,22 @@
 import os
 import json
 
-from todos import decimalencoder
+#from todos import decimalencoder
+import decimalencoder
 import boto3
+from todoTableClass import handler as todoTableClass
 dynamodb = boto3.resource('dynamodb')
 
 
 def get(event, context):
-    table = dynamodb.Table(os.environ['DYNAMODB_TABLE'])
-
-    # fetch todo from the database
-    result = table.get_item(
-        Key={
-            'id': event['pathParameters']['id']
-        }
-    )
+    tdGet = todoTableClass(table = os.environ['DYNAMODB_TABLE'], dynamodb = dynamodb)
+    item = tdGet.get_todo(event['pathParameters']['id'])
 
     # create a response
     response = {
         "statusCode": 200,
-        "body": json.dumps(result['Item'],
-                           cls=decimalencoder.DecimalEncoder)
+        "body": json.dumps(item, cls=decimalencoder.DecimalEncoder)
     }
 
     return response
+