+++ todos/update.py
@@ -3,8 +3,10 @@ import time
 import logging
 import os
 
-from todos import decimalencoder
+#from todos import decimalencoder
+import decimalencoder
 import boto3
+from todoTableClass import handler as todoTableClass
 dynamodb = boto3.resource('dynamodb')
 
 
@@ -15,34 +17,15 @@ def update(event, context):
         raise Exception("Couldn't update the todo item.")
         return
 
-    timestamp = int(time.time() * 1000)
-
-    table = dynamodb.Table(os.environ['DYNAMODB_TABLE'])
+    tdUpdate = todoTableClass(table = os.environ['DYNAMODB_TABLE'], dynamodb = dynamodb)
 
     # update the todo in the database
-    result = table.update_item(
-        Key={
-            'id': event['pathParameters']['id']
-        },
-        ExpressionAttributeNames={
-          '#todo_text': 'text',
-        },
-        ExpressionAttributeValues={
-          ':text': data['text'],
-          ':checked': data['checked'],
-          ':updatedAt': timestamp,
-        },
-        UpdateExpression='SET #todo_text = :text, '
-                         'checked = :checked, '
-                         'updatedAt = :updatedAt',
-        ReturnValues='ALL_NEW',
-    )
+    item = tdUpdate.update_todo(id = event['pathParameters']['id'], text = data['text'], checked = data['checked'])
 
     # create a response
     response = {
         "statusCode": 200,
-        "body": json.dumps(result['Attributes'],
-                           cls=decimalencoder.DecimalEncoder)
+        "body": json.dumps(item['Attributes'], cls=decimalencoder.DecimalEncoder)
     }
 
     return response